SLING-11571 reduce duplication and other cleanup
