+++ src/main/java/org/apache/sling/repoinit/parser/operations/AddMixins.java
@@ -25,13 +25,10 @@ import org.jetbrains.annotations.NotNull;
 import org.osgi.annotation.versioning.ProviderType;
 
 @ProviderType
-public class AddMixins extends Operation {
-    private final List<String> paths;
-    private final List<String> mixins;
+public class AddMixins extends BaseMixinsOperation {
 
     public AddMixins(List<String> mixins, List<String> paths) {
-        this.mixins = mixins;
-        this.paths = paths;
+        super(mixins, paths);
     }
 
     @Override
@@ -42,30 +39,18 @@ public class AddMixins extends Operation {
     @Override
     protected String getParametersDescription() {
         final StringBuilder sb = new StringBuilder();
-        if (mixins != null) {
-            sb.append(String.join(",", mixins));
-        }
-        if (paths != null) {
-            sb.append(" to ").append(String.join(",", paths));
-        }
+        sb.append(listToString(getMixins()));
+        sb.append(" to ").append(pathsToString(getPaths()));
         return sb.toString();
     }
 
     @NotNull
     @Override
     public String asRepoInitString() {
-        // FIXME: see SLING-10238 for type and quoted values that cannot be generated
-        //        exactly as they were originally defined in repo-init
         try (Formatter formatter = new Formatter()) {
-            formatter.format("add mixin %s to %s%n", listToString(mixins), pathsToString(paths));
+            formatter.format("add mixin %s to %s%n", listToString(getMixins()), pathsToString(getPaths()));
             return formatter.toString();
         }
     }
-
-    public List<String> getPaths() {
-        return paths;
-    }
-
-    public List<String> getMixins () {return mixins;}
 
 }
