+++ src/main/java/org/apache/sling/repoinit/parser/operations/BaseMixinsOperation.java
@@ -1,0 +1,39 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.apache.sling.repoinit.parser.operations;
+
+
+import java.util.Collections;
+import java.util.List;
+
+public abstract class BaseMixinsOperation extends Operation {
+    private final List<String> paths;
+    private final List<String> mixins;
+
+    protected BaseMixinsOperation(List<String> mixins, List<String> paths) {
+        this.mixins = mixins != null ? mixins : Collections.emptyList();
+        this.paths = paths != null ? paths : Collections.emptyList();
+    }
+
+    public List<String> getPaths() {
+        return paths;
+    }
+
+    public List<String> getMixins () {return mixins;}
+
+}
